Move test codes to seperated directories
