+++ lib/js/test/Auth/MenuAuth.spec.js
@@ -1,5 +1,5 @@
 import { assert } from 'chai';
-import MenuAuth from './MenuAuth';
+import MenuAuth from '../../lib/Auth/MenuAuth';
 
 describe('MenuAuth', () => {
   it('tests if MenuAuth includes test url', () => {
