+++ lib/js/test/Session/CmsSessionStore.spec.js
@@ -1,5 +1,5 @@
 import { assert } from 'chai';
-import CmsSessionStore from './CmsSessionStore';
+import CmsSessionStore from '../../lib/Session/CmsSessionStore';
 
 describe('CmsSessionStore', () => {
   it('decodes php session', () => {
